@@ -1,8 +1,17 @@
-module github.com/KpnmServer/go-openbmclapi
+module github.com/LiterMC/go-openbmclapi
 
-go 1.15
+go 1.18
 
 require (
 	github.com/KpnmServer/go-util v0.0.0-20210918230454-6dd48583efa9
 	github.com/gorilla/websocket v1.4.2
+)
+
+require (
+	github.com/hamba/avro/v2 v2.18.0 // indirect
+	github.com/json-iterator/go v1.1.12 // indirect
+	github.com/klauspost/compress v1.17.4 // indirect
+	github.com/mitchellh/mapstructure v1.5.0 // indirect
+	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
+	github.com/modern-go/reflect2 v1.0.2 // indirect
 )
