@@ -1,14 +1,13 @@
-
 package main
 
 import (
 	// io "io"
-	os "os"
 	bytes "bytes"
-	strings "strings"
 	fmt "fmt"
-	time "time"
+	os "os"
+	strings "strings"
 	sync "sync"
+	time "time"
 
 	ufile "github.com/KpnmServer/go-util/file"
 )
@@ -19,7 +18,7 @@ var logfile *os.File
 var logLock sync.Mutex
 var logTimeFormat string = "15:04:05.000"
 
-func logX(x string, args ...interface{}){
+func logX(x string, args ...any) {
 	sa := make([]string, len(args))
 	for i, _ := range args {
 		sa[i] = fmt.Sprint(args[i])
@@ -42,7 +41,7 @@ func logX(x string, args ...interface{}){
 	}
 }
 
-func logXf(x string, format string, args ...interface{}){
+func logXf(x string, format string, args ...any) {
 	c := fmt.Sprintf(format, args...)
 	buf := bytes.NewBuffer(nil)
 	buf.Grow(6 + len(logTimeFormat) + len(x) + len(c)) // (1 + len(x) + 2 + len(logTimeFormat) + 3)
@@ -61,48 +60,48 @@ func logXf(x string, format string, args ...interface{}){
 	}
 }
 
-func logDebug(args ...interface{}){
+func logDebug(args ...any) {
 	if DEBUG {
 		logX("DBUG", args...)
 	}
 }
 
-func logDebugf(format string, args ...interface{}){
+func logDebugf(format string, args ...any) {
 	if DEBUG {
 		logXf("DBUG", format, args...)
 	}
 }
 
-func logInfo(args ...interface{}){
+func logInfo(args ...any) {
 	logX("INFO", args...)
 }
 
-func logInfof(format string, args ...interface{}){
+func logInfof(format string, args ...any) {
 	logXf("INFO", format, args...)
 }
 
-func logWarn(args ...interface{}){
+func logWarn(args ...any) {
 	logX("WARN", args...)
 }
 
-func logWarnf(format string, args ...interface{}){
+func logWarnf(format string, args ...any) {
 	logXf("WARN", format, args...)
 }
 
-func logError(args ...interface{}){
+func logError(args ...any) {
 	logX("ERRO", args...)
 }
 
-func logErrorf(format string, args ...interface{}){
+func logErrorf(format string, args ...any) {
 	logXf("ERRO", format, args...)
 }
 
-func flushLogfile(){
-	if ufile.IsNotExist(logdir){
+func flushLogfile() {
+	if ufile.IsNotExist(logdir) {
 		ufile.CreateDir(logdir)
 	}
 	lfile, err := os.OpenFile(ufile.JoinPath(logdir, time.Now().Format("20060102-15.log")),
-		os.O_WRONLY | os.O_APPEND | os.O_CREATE, 0666)
+		os.O_WRONLY|os.O_APPEND|os.O_CREATE, 0666)
 	if err != nil {
 		logError("Create new log file error:", err)
 		return
@@ -112,17 +111,16 @@ func flushLogfile(){
 	logfile = lfile
 }
 
-func init(){
+func init() {
 	flushLogfile()
-	go func(){
-		tma := (time.Now().Unix() / (60 * 60) + 1) * (60 * 60)
-		for{
-			select{
-			case <-time.After(time.Duration(tma - time.Now().Unix()) * time.Second):
-				tma = (time.Now().Unix() / (60 * 60) + 1) * (60 * 60)
+	go func() {
+		tma := (time.Now().Unix()/(60*60) + 1) * (60 * 60)
+		for {
+			select {
+			case <-time.After(time.Duration(tma-time.Now().Unix()) * time.Second):
+				tma = (time.Now().Unix()/(60*60) + 1) * (60 * 60)
 				flushLogfile()
 			}
 		}
 	}()
 }
-
