@@ -1,75 +1,72 @@
-
 package main
 
 import (
-	os "os"
-	io "io"
-	time "time"
-	sort "sort"
-	strings "strings"
-	strconv "strconv"
-	fmt "fmt"
-	sync "sync"
-	context "context"
-	tls "crypto/tls"
-	http "net/http"
+	"context"
+	"crypto"
+	"crypto/tls"
+	"encoding/hex"
+	"fmt"
+	"io"
+	"net/http"
+	"os"
+	"sort"
+	"sync"
+	"sync/atomic"
+	"time"
 
 	ufile "github.com/KpnmServer/go-util/file"
 	json "github.com/KpnmServer/go-util/json"
+	"github.com/hamba/avro/v2"
+	"github.com/klauspost/compress/zstd"
 )
 
-type Cluster struct{
-	host string
-	public_port uint16
-	username string
-	password string
-	version string
-	useragent string
-	prefix string
+type Cluster struct {
+	host       string
+	publicPort uint16
+	username   string
+	password   string
+	version    string
+	useragent  string
+	prefix     string
 
 	cachedir string
-	hits uint32
-	hbytes uint64
-	max_conn uint
-	issync bool
+	hits     atomic.Int32
+	hbytes   atomic.Int64
+	maxConn  int
+	issync   bool
 
-	enabled bool
-	socket *Socket
-	connlock sync.Mutex
-	keepalive func()(bool)
+	mux         sync.Mutex
+	enabled     bool
+	socket      *Socket
+	keepalive   context.CancelFunc
+	downloading map[string]chan struct{}
 
 	client *http.Client
 	Server *http.Server
 }
 
 func NewCluster(
-	host string, public_port uint16,
+	host string, publicPort uint16,
 	username string, password string,
-	version string, address string)(cr *Cluster){
+	version string, address string) (cr *Cluster) {
 	cr = &Cluster{
-		host: host,
-		public_port: public_port,
-		username: username,
-		password: password,
-		version: version,
-		useragent: "openbmclapi-cluster/" + version,
-		prefix: "https://openbmclapi.bangbang93.com",
+		host:       host,
+		publicPort: publicPort,
+		username:   username,
+		password:   password,
+		version:    version,
+		useragent:  "openbmclapi-cluster/" + version,
+		prefix:     "https://openbmclapi.bangbang93.com",
 
 		cachedir: "cache",
-		hits: 0,
-		hbytes: 0,
-		max_conn: 400,
-		issync: false,
-
-		enabled: false,
-		socket: nil,
-		connlock: sync.Mutex{},
+		maxConn:  400,
+		issync:   false,
 
 		client: &http.Client{
-			Timeout: time.Minute * 60,
+			Timeout: time.Second * 60,
 			Transport: &http.Transport{
 				TLSClientConfig: &tls.Config{
-					InsecureSkipVerify: true, // Skip verify because the author was lazy
+					// InsecureSkipVerify: true, // Skip verify because the author was lazy
 				},
 			},
 		},
@@ -81,15 +78,11 @@ func NewCluster(
 	return
 }
 
-func (cr *Cluster)Enable()(bool){
+func (cr *Cluster) Enable() bool {
+	cr.mux.Lock()
+	defer cr.mux.Unlock()
 	if cr.socket != nil {
-		logDebug("Extara enable")
-		return true
-	}
-	cr.connlock.Lock()
-	defer cr.connlock.Unlock()
-	if cr.socket != nil {
-		logDebug("Extara after lock enable")
+		logDebug("Extra enable")
 		return true
 	}
 	var (
@@ -104,16 +97,19 @@ func (cr *Cluster)Enable()(bool){
 	header.Set("Origin", cr.prefix)
 
 	cr.socket = NewSocket(NewESocket())
-	cr.socket.ConnectHandle = func(*Socket){
+	cr.socket.ConnectHandle = func(*Socket) {
 		cr._enable()
 	}
-	cr.socket.DisconnectHandle = func(*Socket){
+	cr.socket.DisconnectHandle = func(*Socket) {
 		cr.Disable()
 	}
-	cr.socket.GetIO().ErrorHandle = func(*ESocket){
-		go func(){
+	cr.socket.GetIO().ErrorHandle = func(*ESocket) {
+		go func() {
 			cr.enabled = false
-			cr.Disable()()
+			ch := cr.Disable()
+			if ch != nil {
+				<-ch
+			}
 			if !cr.Enable() {
 				logError("Cannot reconnect to server, exit.")
 				panic("Cannot reconnect to server, exit.")
@@ -128,8 +124,8 @@ func (cr *Cluster)Enable()(bool){
 	return true
 }
 
-func (cr *Cluster)_enable(){
-	cr.socket.EmitAck(func(_ uint64, data json.JsonArr){
+func (cr *Cluster) _enable() {
+	cr.socket.EmitAck(func(_ uint64, data json.JsonArr) {
 		data = data.GetArray(0)
 		if len(data) < 2 || !data.GetBool(1) {
 			logError("Enable failed: " + data.String())
@@ -137,28 +133,26 @@ func (cr *Cluster)_enable(){
 		}
 		logInfo("get enable ack:", data)
 		cr.enabled = true
-		cr.keepalive = createInterval(func(){
+
+		var keepaliveCtx context.Context
+		keepaliveCtx, cr.keepalive = context.WithCancel(context.TODO())
+		createInterval(keepaliveCtx, func() {
 			cr.KeepAlive()
-		}, time.Second * 60)
+		}, KeepAliveInterval)
 	}, "enable", json.JsonObj{
-		"host": cr.host,
-		"port": cr.public_port,
+		"host":    cr.host,
+		"port":    cr.publicPort,
 		"version": cr.version,
 	})
 }
 
-func (cr *Cluster)KeepAlive()(ok bool){
-	hits, hbytes := cr.hits, cr.hbytes
-	var (
-		err error
-	)
-	err = cr.socket.EmitAck(func(_ uint64, data json.JsonArr){
+func (cr *Cluster) KeepAlive() (ok bool) {
+	hits, hbytes := cr.hits.Swap(0), cr.hbytes.Swap(0)
+	err := cr.socket.EmitAck(func(_ uint64, data json.JsonArr) {
 		data = data.GetArray(0)
 		if len(data) > 1 && data.Get(0) == nil {
-			logInfo("Keep-alive success:", hits, bytesToUnit((float32)(hbytes)), data)
-			cr.hits -= hits
-			cr.hbytes -= hbytes
-		}else{
+			logInfo("Keep-alive success:", hits, bytesToUnit((float64)(hbytes)), data)
+		} else {
 			logInfo("Keep-alive failed:", data.Get(0))
 			cr.Disable()
 			if !cr.Enable() {
@@ -167,53 +161,58 @@ func (cr *Cluster)KeepAlive()(ok bool){
 			}
 		}
 	}, "keep-alive", json.JsonObj{
-		"time": time.Now().UTC().Format("2006-01-02T15:04:05Z"),
-		"hits": cr.hits,
-		"bytes": cr.hbytes,
+		"time":  time.Now().UTC().Format("2006-01-02T15:04:05Z"),
+		"hits":  hits,
+		"bytes": hbytes,
 	})
 	if err != nil {
 		logError("Error when keep-alive:", err)
+		return false
 	}
-	return err == nil
+	return true
 }
 
-func (cr *Cluster)Disable()(sync func()){
+func (cr *Cluster) Disable() (done <-chan struct{}) {
+	cr.mux.Lock()
+	defer cr.mux.Unlock()
+
 	if !cr.enabled {
-		logDebug("Extara disable")
-		return func(){}
+		logDebug("Extra disable")
+		return nil
 	}
 	logInfo("Disabling cluster")
 	if cr.keepalive != nil {
 		cr.keepalive()
 		cr.keepalive = nil
 	}
-	if cr.socket != nil {
-		sch := make(chan struct{}, 1)
-		err := cr.socket.EmitAck(func(_ uint64, data json.JsonArr){
-			data = data.GetArray(0)
-			logInfo("disable ack:", data)
-			if !data.GetBool(1) {
-				logError("Disable failed: " + data.String())
-				panic("Disable failed: " + data.String())
-			}
-			sch <- struct{}{}
-		}, "disable")
-		cr.enabled = false
-		cr.socket.Close()
-		cr.socket = nil
-		if err == nil {
-			return func(){
-				<- sch
-			}
-		}
+	if cr.socket == nil {
+		return nil
 	}
-	return func(){}
+	sch := make(chan struct{}, 1)
+	err := cr.socket.EmitAck(func(_ uint64, data json.JsonArr) {
+		data = data.GetArray(0)
+		logInfo("disable ack:", data)
+		if !data.GetBool(1) {
+			logError("Disable failed: " + data.String())
+			panic("Disable failed: " + data.String())
+		}
+		sch <- struct{}{}
+	}, "disable")
+	cr.enabled = false
+	cr.socket.Close()
+	cr.socket = nil
+	if err != nil {
+		return nil
+	}
+	return sch
 }
 
-func (cr *Cluster)queryFunc(method string, url string, call func(*http.Request))(res *http.Response, err error){
+func (cr *Cluster) queryFunc(method string, url string, call func(*http.Request)) (res *http.Response, err error) {
 	var req *http.Request
-	req, err = http.NewRequest(method, cr.prefix + url, nil)
-	if err != nil { return }
+	req, err = http.NewRequest(method, cr.prefix+url, nil)
+	if err != nil {
+		return
+	}
 	req.SetBasicAuth(cr.username, cr.password)
 	req.Header.Set("User-Agent", cr.useragent)
 	if call != nil {
@@ -223,12 +222,12 @@ func (cr *Cluster)queryFunc(method string, url string, call func(*http.Request))
 	return
 }
 
-func (cr *Cluster)queryURL(method string, url string)(res *http.Response, err error){
+func (cr *Cluster) queryURL(method string, url string) (res *http.Response, err error) {
 	return cr.queryFunc(method, url, nil)
 }
 
-func (cr *Cluster)queryURLHeader(method string, url string, header map[string]string)(res *http.Response, err error){
-	return cr.queryFunc(method, url, func(req *http.Request){
+func (cr *Cluster) queryURLHeader(method string, url string, header map[string]string) (res *http.Response, err error) {
+	return cr.queryFunc(method, url, func(req *http.Request) {
 		if header != nil {
 			for k, v := range header {
 				req.Header.Set(k, v)
@@ -237,22 +236,31 @@ func (cr *Cluster)queryURLHeader(method string, url string, header map[string]st
 	})
 }
 
-type FileInfo struct{
-	Path string `json:"path"`
-	Hash string `json:"hash"`
-	Size int64  `json:"size"`
-}
-
-type IFileList struct{
-	Files []FileInfo `json:"files"`
-}
-
-func (cr *Cluster)getHashPath(hash string)(string){
+func (cr *Cluster) getHashPath(hash string) string {
 	return ufile.JoinPath(cr.cachedir, hash[:2], hash)
 }
 
-func (cr *Cluster)GetFileList()(files []FileInfo){
-	var(
+type FileInfo struct {
+	Path string `json:"path" avro:"path"`
+	Hash string `json:"hash" avro:"hash"`
+	Size int64  `json:"size" avro:"size"`
+}
+
+// from <https://github.com/bangbang93/openbmclapi/blob/master/src/cluster.ts>
+var fileListSchema = avro.MustParse(`{
+  "type": "array",
+  "items": {
+    "type": "record",
+    "fields": [
+      {"name": "path", "type": "string"},
+      {"name": "hash", "type": "string"},
+      {"name": "size", "type": "long"},
+    ]
+  }
+}`)
+
+func (cr *Cluster) GetFileList() (files []FileInfo) {
+	var (
 		err error
 		res *http.Response
 	)
@@ -261,210 +269,100 @@ func (cr *Cluster)GetFileList()(files []FileInfo){
 		logError("Query filelist error:", err)
 		return nil
 	}
-	list := new(IFileList)
-	err = json.ReadJson(res.Body, &list)
+	defer res.Body.Close()
+	if res.StatusCode != http.StatusOK {
+		logErrorf("Query filelist error: Unexpected status code: %d %s", res.StatusCode, res.Status)
+		return nil
+	}
+	zr, err := zstd.NewReader(res.Body)
 	if err != nil {
 		logError("Parse filelist body error:", err)
 		return nil
 	}
-	return list.Files
+	defer zr.Close() // TODO: reuse the decoder?
+	if err = avro.NewDecoderForSchema(fileListSchema, zr).Decode(&files); err != nil {
+		logError("Parse filelist body error:", err)
+		return nil
+	}
+	return files
 }
 
-type extFileInfo struct{
+type extFileInfo struct {
 	*FileInfo
-	dlerr error
+	dlerr    error
 	trycount int
 }
 
-func (cr *Cluster)SyncFiles(_files []FileInfo, _ctx ...context.Context){
-	var ctx context.Context = nil
-	if len(_ctx) > 0 {
-		ctx = _ctx[0]
-	}
-	if checkContext(ctx) {
-		logWarn("File sync interrupted")
-		return
-	}
+type syncStats struct {
+	totalsize  float64
+	downloaded float64
+	slots      chan struct{}
+	fcount     atomic.Int32
+	fl         int
+}
+
+func (cr *Cluster) SyncFiles(files0 []FileInfo, ctx context.Context) {
 	logInfo("Pre sync files...")
 	if cr.issync {
 		logWarn("Another sync task is running!")
 		return
 	}
 	cr.issync = true
-	defer func(){
+	defer func() {
 		cr.issync = false
 	}()
+
 	var files []FileInfo
-	if !withContext(ctx, func(){
-		files = cr.CheckFiles(_files, make([]FileInfo, 0, 16))
+	if !withContext(ctx, func() {
+		files = cr.CheckFiles(files0, make([]FileInfo, 0, 16))
 	}) {
 		logWarn("File sync interrupted")
 		return
 	}
+
 	fl := len(files)
 	if fl == 0 {
 		logInfo("All file was synchronized")
-		go cr.gc(_files)
+		go cr.gc(files0)
 		return
 	}
-	sort.Slice(files, func(i, j int)(bool){ return files[i].Size > files[j].Size })
-	var (
-		totalsize float32 = 0
-		downloaded float32 = 0
-	)
-	for i, _ := range files { totalsize += (float32)(files[i].Size) }
-	logInfof("Starting sync files, count: %d, total: %s", fl, bytesToUnit(totalsize))
-	start := time.Now()
-	re := make(chan *extFileInfo, (int)(cr.max_conn))
-	fcount := 0
-	alive := (uint)(0)
-	var (
-		dlhandle func(f *extFileInfo, c chan<- *extFileInfo, ctx context.Context)
-		handlef func(f *extFileInfo, ctx context.Context)
-		dlfile func(f *extFileInfo, ctx context.Context)
-	)
-	dlhandle = func(f *extFileInfo, c chan<- *extFileInfo, ctx context.Context){
-		defer func(){
-			alive--
-			c <- f
-		}()
-		var(
-			buf []byte
-			bi int
-			n int
-			err error
-			res *http.Response
-			fd *os.File
-		)
-		buf, bi = mallocBuf()
-		defer func(){
-			releaseBuf(bi)
-		}()
-		p := cr.getHashPath(f.Hash)
-		defer func(){
-			f.dlerr = err
-			if err != nil {
-				f.trycount++
-				if ufile.IsExist(p) {
-					os.Remove(p)
-				}
-			}
-		}()
-		logDebug("Downloading:", f.Path)
-		for i := 0; i < 3 ;i++ {
-			if checkContext(ctx) { return }
-			res, err = cr.queryURL("GET", f.Path)
-			if err != nil {
-				continue
-			}
-			defer res.Body.Close()
-			if checkContext(ctx) { return }
-			if res.StatusCode != http.StatusOK {
-				err = fmt.Errorf("Unexpected status code: %d", res.StatusCode)
-				return
-			}
-			fd, err = os.Create(p)
-			if err != nil {
-				continue
-			}
-			defer fd.Close()
-			var t int64 = 0
-			for {
-				if checkContext(ctx) { return }
-				n, err = res.Body.Read(buf)
-				if n == 0 {
-					if err == io.EOF{
-						err = nil
-					}
-					break
-				}
-				t += (int64)(n)
-				_, err = fd.Write(buf[:n])
-				if err != nil { break }
-			}
-			if t != f.Size && err == nil {
-				err = fmt.Errorf("File size wrong %s expect %s", bytesToUnit((float32)(t)), bytesToUnit((float32)(f.Size)))
-				if DEBUG {
-					f0, _ := os.Open(p)
-					b0, _ := io.ReadAll(f0)
-					if len(b0) < 16 * 1024 {
-						logDebug("File content:", (string)(b0), "//for", f.Path)
-					}
-				}
-			}
-			if err != nil {
-				fd.Close()
-				continue
-			}
-			break
-		}
-	}
-	handlef = func(f *extFileInfo, ctx context.Context){
-		if checkContext(ctx) { return }
-		if f.dlerr != nil {
-			logErrorf("Download file error: %v %s [%s/%s;%d/%d]%.2f%%", f.dlerr, f.Path,
-				bytesToUnit(downloaded), bytesToUnit(totalsize),
-				fcount, fl, downloaded / totalsize * 100)
-			if f.trycount < 1 { // ps: 由于主服务器数据库维护中, 故此减小max_trycount, 以提高初次启动速度(另:该方法治标不治本
-				c, _ := context.WithCancel(ctx)
-				dlfile(f, c)
-			}else{
-				fcount++
-			}
-		}else{
-			downloaded += (float32)(f.Size)
-			fcount++
-			logInfof("Downloaded: %s [%s/%s;%d/%d]%.2f%%", f.Path,
-				bytesToUnit(downloaded), bytesToUnit(totalsize),
-				fcount, fl, downloaded / totalsize * 100)
-		}
-	}
-	dlfile = func(f *extFileInfo, ctx context.Context){
-		for alive >= cr.max_conn {
-			select{
-			case r := <-re:
-				c, _ := context.WithCancel(ctx)
-				handlef(r, c)
-			case <-ctx.Done():
-				return
-			}
-		}
-		alive++
-		go dlhandle(f, re, ctx)
-	}
+
+	// sort the files in descending order of size
+	sort.Slice(files, func(i, j int) bool { return files[i].Size > files[j].Size })
+
+	var stats syncStats
+	stats.slots = make(chan struct{}, cr.maxConn)
+	stats.fl = fl
 	for i, _ := range files {
-		if checkContext(ctx) {
-			logWarn("File sync interrupted")
-			return
-		}
-		c, _ := context.WithCancel(ctx)
-		dlfile(&extFileInfo{ FileInfo: &files[i], dlerr: nil, trycount: 0 }, c)
+		stats.totalsize += (float64)(files[i].Size)
 	}
-	for fcount < fl {
-		select{
-		case r := <-re:
-			c, _ := context.WithCancel(ctx)
-			handlef(r, c)
+
+	logInfof("Starting sync files, count: %d, total: %s", fl, bytesToUnit(stats.totalsize))
+	start := time.Now()
+
+	for i, _ := range files {
+		cr.dlfile(ctx, &stats, &extFileInfo{FileInfo: &files[i], dlerr: nil, trycount: 0})
+	}
+	for i := len(stats.slots); i > 0; i-- {
+		select {
+		case stats.slots <- struct{}{}:
 		case <-ctx.Done():
 			logWarn("File sync interrupted")
 			return
 		}
 	}
+
 	use := time.Since(start)
-	logInfof("All file was synchronized, use time: %v, %s/s", use, bytesToUnit(totalsize / (float32)(use / time.Second)))
+	logInfof("All file was synchronized, use time: %v, %s/s", use, bytesToUnit(stats.totalsize/use.Seconds()))
 	cr.issync = false
 	var flag bool = false
-	if use > time.Minute * 10 { // interval time
-		logWarn("Synchronization time was more than 10 min, re checking now.")
-		_files2 := cr.GetFileList()
-		if len(_files2) != len(_files) {
+	if use > SyncFileInterval {
+		logWarn("Synchronization time was more than 10 min, re-check now.")
+		files2 := cr.GetFileList()
+		if len(files2) != len(files0) {
 			flag = true
-		}else{
-			for _, f := range _files2 {
-				if checkContext(ctx){
-					logWarn("File sync interrupted")
-					return
-				}
+		} else {
+			for _, f := range files2 {
 				p := cr.getHashPath(f.Hash)
 				if ufile.IsNotExist(p) {
 					flag = true
@@ -474,14 +372,76 @@ func (cr *Cluster)SyncFiles(_files []FileInfo, _ctx ...context.Context){
 		}
 		if flag {
 			logWarn("At least one file has changed during file synchronization, re synchronize now.")
-			cr.SyncFiles(_files2, ctx)
+			cr.SyncFiles(files2, ctx)
 			return
 		}
 	}
-	go cr.gc(_files)
+	go cr.gc(files0)
 }
 
-func (cr *Cluster)CheckFiles(files []FileInfo, notf []FileInfo)([]FileInfo){
+func (cr *Cluster) dlfile(ctx context.Context, stats *syncStats, f *extFileInfo) {
+WAIT_SLOT:
+	for {
+		select {
+		case stats.slots <- struct{}{}:
+			break WAIT_SLOT
+		case <-ctx.Done():
+			return
+		}
+	}
+	go func() {
+		defer func() {
+			<-stats.slots
+		}()
+	RETRY:
+		err := cr.dlhandle(ctx, f.FileInfo)
+		if err != nil {
+			logErrorf("Download file error: %v %s [%s/%s;%d/%d]%.2f%%",
+				err, f.Path,
+				bytesToUnit(stats.downloaded), bytesToUnit(stats.totalsize),
+				stats.fcount.Load(), stats.fl,
+				stats.downloaded/stats.totalsize*100)
+			if f.trycount < 3 {
+				f.trycount++
+				goto RETRY
+			}
+			stats.fcount.Add(1)
+			p := cr.getHashPath(f.Hash)
+			if ufile.IsExist(p) {
+				os.Remove(p)
+			}
+		} else {
+			stats.downloaded += (float64)(f.Size)
+			stats.fcount.Add(1)
+			logInfof("Downloaded: %s [%s/%s;%d/%d]%.2f%%", f.Path,
+				bytesToUnit(stats.downloaded), bytesToUnit(stats.totalsize),
+				stats.fcount.Load(), stats.fl,
+				stats.downloaded/stats.totalsize*100)
+		}
+	}()
+}
+
+func (cr *Cluster) dlhandle(ctx context.Context, f *FileInfo) (err error) {
+	logDebug("Downloading:", f.Path)
+	hashMethod, err := getHashMethod(len(f.Hash))
+	if err != nil {
+		return
+	}
+
+	var buf []byte
+	{
+		buf0 := bufPool.Get().(*[]byte)
+		defer bufPool.Put(buf0)
+		buf = *buf0
+	}
+
+	for i := 0; i < 3; i++ {
+		cr.downloadFileBuf(ctx, f, hashMethod, buf)
+	}
+	return
+}
+
+func (cr *Cluster) CheckFiles(files []FileInfo, notf []FileInfo) []FileInfo {
 	logInfo("Starting check files")
 	for i, _ := range files {
 		if cr.issync && notf == nil {
@@ -493,23 +453,23 @@ func (cr *Cluster)CheckFiles(files []FileInfo, notf []FileInfo)([]FileInfo){
 		if err == nil {
 			if fs.Size() != files[i].Size {
 				logInfof("Found wrong size file: '%s'(%s) expect %s",
-					p, bytesToUnit((float32)(fs.Size())), bytesToUnit((float32)(files[i].Size)))
+					p, bytesToUnit((float64)(fs.Size())), bytesToUnit((float64)(files[i].Size)))
 				if notf == nil {
 					os.Remove(p)
-				}else{
+				} else {
 					notf = append(notf, files[i])
 				}
 			}
-		}else{
+		} else {
 			if notf != nil {
 				notf = append(notf, files[i])
 			}
 			if os.IsNotExist(err) {
 				p = ufile.DirPath(p)
 				if ufile.IsNotExist(p) {
-					os.MkdirAll(p, 0744)
+					os.MkdirAll(p, 0755)
 				}
-			}else{
+			} else {
 				os.Remove(p)
 			}
 		}
@@ -518,25 +478,18 @@ func (cr *Cluster)CheckFiles(files []FileInfo, notf []FileInfo)([]FileInfo){
 	return notf
 }
 
-func (cr *Cluster)gc(files []FileInfo){
-	logInfo("Starting global cleanup")
-	fileset := make(map[string]struct{})
+func (cr *Cluster) gc(files []FileInfo) {
+	logInfo("Starting garbage collector")
+	fileset := make(map[string]struct{}, 128)
 	for i, _ := range files {
 		fileset[cr.getHashPath(files[i].Hash)] = struct{}{}
 	}
-	stack := make([]string, 0, 1)
+	stack := make([]string, 0, 10)
 	stack = append(stack, cr.cachedir)
-	var (
-		ok bool
-		p string
-		n string
-		fil []os.DirEntry
-		err error
-	)
 	for len(stack) > 0 {
-		p = stack[len(stack) - 1]
-		stack = stack[:len(stack) - 1]
-		fil, err = os.ReadDir(p)
+		p := stack[len(stack)-1]
+		stack = stack[:len(stack)-1]
+		fil, err := os.ReadDir(p)
 		if err != nil {
 			continue
 		}
@@ -545,120 +498,77 @@ func (cr *Cluster)gc(files []FileInfo){
 				logWarn("Global cleanup interrupted")
 				return
 			}
-			n = ufile.JoinPath(p, f.Name())
+			n := ufile.JoinPath(p, f.Name())
 			if ufile.IsDir(n) {
 				stack = append(stack, n)
-			}else if _, ok = fileset[n]; !ok {
+			} else if _, ok := fileset[n]; !ok {
 				logInfo("Found outdated file:", n)
 				os.Remove(n)
 			}
 		}
 	}
-	logInfo("Global cleanup finished")
+	logInfo("Garbage collector finished")
 }
 
-func (cr *Cluster)DownloadFile(hash string)(bool){
-	var(
-		err error
+func (cr *Cluster) downloadFileBuf(ctx context.Context, f *FileInfo, hashMethod crypto.Hash, buf []byte) (err error) {
+	var (
 		res *http.Response
-		fd *os.File
+		fd  *os.File
 	)
-	res, err = cr.queryURL("GET", "/openbmclapi/download/" + hash + "?noopen=1")
-	if err != nil {
-		logError("Query file error:", err)
-		return false
+	if res, err = cr.queryURL("GET", f.Path); err != nil {
+		return
 	}
-	fd, err = os.Create(cr.getHashPath(hash))
-	if err != nil {
-		logError("Create file error:", err)
-		return false
+	defer res.Body.Close()
+	if err = ctx.Err(); err != nil {
+		return
 	}
-	defer fd.Close()
-	_, err = io.Copy(fd, res.Body)
-	if err != nil {
-		logError("Write file error:", err)
-		return false
+	if res.StatusCode != http.StatusOK {
+		return fmt.Errorf("Unexpected status code: %d", res.StatusCode)
 	}
-	return true
-}
 
-func (cr *Cluster)ServeHTTP(response http.ResponseWriter, request *http.Request){
-	method := request.Method
-	url := request.URL
-	rawpath := url.EscapedPath()
-	if SHOW_SERVE_INFO {
-		go logInfo("serve url:", url.String())
+	hw := hashMethod.New()
+
+	if fd, err = os.Create(cr.getHashPath(f.Hash)); err != nil {
+		return
 	}
-	switch{
-	case strings.HasPrefix(rawpath, "/download/"):
-		if method == "GET" {
-			path := cr.getHashPath(rawpath[10:])
-			if ufile.IsNotExist(path) {
-				if !cr.DownloadFile(path) {
-					response.WriteHeader(http.StatusInternalServerError)
-					return
-				}
-			}
-			if name := request.Form.Get("name"); name != "" {
-				response.Header().Set("Content-Disposition", "attachment; filename=" + name)
-			}
-			response.Header().Set("Cache-Control", "max-age=2592000") // 30 days
-			fd, err := os.Open(path)
-			if err != nil {
-				response.WriteHeader(http.StatusInternalServerError)
-				return
-			}
-			response.WriteHeader(http.StatusOK)
-			buf, bi := mallocBuf()
-			defer func(){
-				releaseBuf(bi)
-			}()
-			var (
-				hb uint64
-				n int
-			)
-			for {
-				n, err = fd.Read(buf)
-				if err != nil {
-					if err == io.EOF {
-						err = nil
-						break
-					}
-					logError("Error when serving download read file:", err)
-					return
-				}
-				if n == 0 {
-					break
-				}
-				_, err = response.Write(buf[:n])
-				if err != nil {
-					if !IGNORE_SERVE_ERROR {
-						logError("Error when serving download:", err)
-					}
-					return
-				}
-				hb += (uint64)(n)
-			}
-			cr.hits++
-			cr.hbytes += hb
-			return
-		}
-	case strings.HasPrefix(rawpath, "/measure/"):
-		if method == "GET"{
-			if request.Header.Get("x-openbmclapi-secret") != cr.password {
-				response.WriteHeader(http.StatusForbidden)
-				return
-			}
-			n, e := strconv.Atoi(rawpath[9:])
-			if e != nil || n < 0 || n > 200 {
-				response.WriteHeader(http.StatusBadRequest)
-				return
-			}
-			response.WriteHeader(http.StatusOK)
-			response.Write(make([]byte, n * 1024 * 1024))
-			return
+
+	var t int64
+	t, err = io.CopyBuffer(io.MultiWriter(hw, fd), res.Body, buf)
+	fd.Close()
+	if err != nil {
+		return
+	}
+	if f.Size >= 0 && t != f.Size {
+		err = fmt.Errorf("File size wrong, got %s, expect %s", bytesToUnit((float64)(t)), bytesToUnit((float64)(f.Size)))
+	} else if hs := hex.EncodeToString(hw.Sum(buf[:0])); hs != f.Hash {
+		err = fmt.Errorf("File hash not match, got %s, expect %s", hs, f.Hash)
+	}
+	if DEBUG && err != nil {
+		f0, _ := os.Open(cr.getHashPath(f.Hash))
+		b0, _ := io.ReadAll(f0)
+		if len(b0) < 16*1024 {
+			logDebug("File content:", (string)(b0), "//for", f.Path)
 		}
 	}
-	response.WriteHeader(http.StatusNotFound)
+	return
 }
 
+func (cr *Cluster) DownloadFile(ctx context.Context, hash string) (err error) {
+	hashMethod, err := getHashMethod(len(hash))
+	if err != nil {
+		return
+	}
+
+	var buf []byte
+	{
+		buf0 := bufPool.Get().(*[]byte)
+		defer bufPool.Put(buf0)
+		buf = *buf0
+	}
+	f := &FileInfo{
+		Path: "/openbmclapi/download/" + hash + "?noopen=1",
+		Hash: hash,
+		Size: -1,
+	}
+	return cr.downloadFileBuf(ctx, f, hashMethod, buf)
+}
