@@ -1,40 +1,32 @@
-
 package main
 
 import (
-	fmt "fmt"
-	time "time"
-	strings "strings"
-	context "context"
+	"context"
+	"crypto"
+	"fmt"
+	"strings"
+	"sync"
+	"time"
 
 	ufile "github.com/KpnmServer/go-util/file"
 )
 
-func hashToFilename(hash string)(string){
+func hashToFilename(hash string) string {
 	return ufile.JoinPath(hash[0:2], hash)
 }
 
-func createInterval(do func(), delay time.Duration)(cancel func()(bool)){
-	running := false
-	exitch := make(chan struct{}, 1)
-	cancel = func()(bool) {
-		if !running {
-			return false
-		}
-		logDebug("Canceled interval:", exitch)
-		running = false
-		exitch <- struct{}{}
-		exitch = nil
-		return true
-	}
-	logDebug("Created interval:", exitch)
-	go func(){
-		running = true
-		for running {
-			select{
-			case <-exitch:
+func createInterval(ctx context.Context, do func(), delay time.Duration) {
+	logDebug("Interval created:", ctx)
+	go func() {
+		ticker := time.NewTicker(delay)
+		defer ticker.Stop()
+
+		for {
+			select {
+			case <-ctx.Done():
+				logDebug("Interval stopped:", ctx)
 				return
-			case <-time.After(delay):
+			case <-ticker.C:
 				do()
 			}
 		}
@@ -42,14 +34,14 @@ func createInterval(do func(), delay time.Duration)(cancel func()(bool)){
 	return
 }
 
-func httpToWs(origin string)(string){
+func httpToWs(origin string) string {
 	if strings.HasPrefix(origin, "http") {
 		return "ws" + origin[4:]
 	}
 	return origin
 }
 
-func bytesToUnit(size float32)(string){
+func bytesToUnit(size float64) string {
 	unit := "Byte"
 	if size >= 1000 {
 		size /= 1024
@@ -70,17 +62,17 @@ func bytesToUnit(size float32)(string){
 	return fmt.Sprintf("%.1f", size) + unit
 }
 
-func withContext(ctx context.Context, call func())(bool){
+func withContext(ctx context.Context, call func()) bool {
 	if ctx == nil {
 		call()
 		return true
 	}
-	done := make(chan struct{}, 1)
-	go func(){
+	done := make(chan struct{}, 0)
+	go func() {
+		defer close(done)
 		call()
-		done <- struct{}{}
 	}()
-	select{
+	select {
 	case <-ctx.Done():
 		return false
 	case <-done:
@@ -88,39 +80,22 @@ func withContext(ctx context.Context, call func())(bool){
 	}
 }
 
-func checkContext(ctx context.Context)(bool){
-	if ctx == nil {
-		return false
-	}
-	select{
-	case <-ctx.Done():
-		return true
-	default:
-		return false
-	}
-}
-
 const BUF_SIZE = 1024 * 512 // 512KB
-const BUF_COUNT = 1024 // 512MB
-var BUFFER_PAGE = make([][]byte, BUF_COUNT)
-var buf_trigger = make(chan int, BUF_COUNT)
-func init(){
-	for i, _ := range BUFFER_PAGE {
-		BUFFER_PAGE[i] = make([]byte, BUF_SIZE)
-		buf_trigger <- i
-	}
+var bufPool = sync.Pool{
+	New: func() any {
+		buf := make([]byte, BUF_SIZE)
+		return &buf
+	},
 }
 
-func mallocBuf()(buf []byte, i int){
-	logDebug("Waiting free buf")
-	i = <- buf_trigger
-	logDebug("Got free buf", i)
-	buf = BUFFER_PAGE[i]
+func getHashMethod(l int) (hashMethod crypto.Hash, err error) {
+	switch l {
+	case 32:
+		hashMethod = crypto.MD5
+	case 40:
+		hashMethod = crypto.SHA1
+	default:
+		err = fmt.Errorf("Unknown hash length %d", l)
+	}
 	return
 }
-
-func releaseBuf(i int){
-	logDebug("Release buf", i)
-	buf_trigger <- i
-}
-
